Move files to library, update references
