+++ lib/terminal.js
