+++ lib/utilities.js
