+++ lib/website.js
@@ -1,7 +1,7 @@
 import * as commands from '../bin/include.js';
-import * as utilities from '../sbin/utilities.js';
-import { User } from '../sbin/user.js';
-import { Terminal } from '../sbin/terminal.js';
+import * as utilities from '../lib/utilities.js';
+import { User } from '../lib/user.js';
+import { Terminal } from '../lib/terminal.js';
 
 $('.disabled').click(function(e){
     e.preventDefault();
